Move create_person, list_people, create_teacher and create_student funcs to personStorage class
